+++ R/ci-ri.R
@@ -13,8 +13,9 @@
 #' @template datasetParam
 #' @template treeParam
 #' 
-#' @return `Consistency()` returns a named vector specifying the consistency index (ci),
-#' retention index (ri), and rescaled consistency index (rc).
+#' @return `Consistency()` returns a named vector specifying the 
+#' consistency index (`ci`),
+#' retention index (`ri`), and rescaled consistency index (`rc`).
 #' 
 #' @examples 
 #' data(inapplicable.datasets)